refactor(useListbox): useActiveOptionを統合してテストを移行
useActiveOptionはuseListboxからのみ使用されていたため、useListboxに統合した。
これにより、二段階の依存配列管理が不要になり、コードがシンプルになった。

useActiveOptionのテスト内容は、useListbox.activeOption.test.tsxに移行し、
キーボードイベント経由でactiveOptionの動作を検証するように変更した。
